+++ packages/components/organisms/f-checkout/src/store/checkout.module.js
@@ -7,6 +7,8 @@ import {
     UPDATE_CUSTOMER_DETAILS,
     UPDATE_FULFILMENT_ADDRESS,
     UPDATE_FULFILMENT_TIME,
+    UPDATE_IS_FULFILLABLE,
+    UPDATE_ISSUES,
     UPDATE_STATE,
     UPDATE_USER_NOTE
 } from './mutation-types';
@@ -35,6 +37,7 @@ export default {
         },
         userNote: '',
         isFulfillable: true,
+        issues: [],
         notices: [],
         messages: [],
         availableFulfilment: {
@@ -78,7 +81,7 @@ export default {
          * @param {Object} payload - Parameter with the different configurations for the request.
          */
         // eslint-disable-next-line no-unused-vars
-        patchCheckout: async ({ state }, {
+        patchCheckout: async ({ commit, state }, {
             url, data, timeout
         }) => {
             // TODO: deal with exceptions and handle this action properly (when the functionality is ready)
@@ -94,8 +97,12 @@ export default {
                 timeout
             };
 
-            // eslint-disable-next-line no-unused-vars
-            const response = await axios.patch(url, data, config);
+            // TODO - Handle and log any errors
+            const { data: responseData } = await axios.patch(url, data, config);
+            const { issues, isFulfillable } = responseData;
+
+            commit(UPDATE_IS_FULFILLABLE, isFulfillable);
+            commit(UPDATE_ISSUES, issues);
         },
 
         /**
@@ -264,6 +271,14 @@ export default {
                 ...state.time,
                 ...time
             };
+        },
+
+        [UPDATE_IS_FULFILLABLE]: (state, isFulfillable) => {
+            state.isFulfillable = isFulfillable;
+        },
+
+        [UPDATE_ISSUES]: (state, issues) => {
+            state.issues = issues;
         },
 
         [UPDATE_USER_NOTE]: (state, userNote) => {