+++ packages/components/molecules/f-searchbox/CHANGELOG.md
@@ -3,6 +3,16 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+v4.0.0-beta.23
+------------------------------
+*February 4, 2021*
+
+### Added
+- Clear button to Loqate feature so addresses can be cleared.
+- Refactor loading indicator component so it's shared.
+- Added tests.
+
+
 v4.0.0-beta.22
 ------------------------------
 *February 2, 2021*